requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
redis>=4.2.0,<5.0.0
fastapi-cache2[redis]==0.2.2
aiofiles>=24.1.0
